Flytt state ut av router
Flytter tilstander ut av routeren inn i den utenforliggende Organisasjonsdetaljer-provideren.
Det skal ikke være gjort noen endringer i logikken.
